style: renamed exit code file
renamed exit code file and removed duplicate line in readme
