+++ .gitignore
@@ -6,8 +6,7 @@
 ##############################
 ## tobot                    ##
 ##############################
-.[Cc][Aa][Cc][Hh][Ee]/
-logs
+[Ll][Oo][Gg][Ss]/
 
 ##############################
 ## clang                    ##
